backend/pyproject.toml
[tool.poetry]
name = "project-management-system"
version = "0.1.0"
description = "A multi-tenant project management system."
authors = ["Cascade AI"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.109.2"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.29"}
psycopg-binary = "^3.1.18"
alembic = "^1.13.1"
pydantic = {extras = ["email"], version = "^2.11.7"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-dotenv = "^1.0.1"
pydantic-settings = "^2.2.1"
asyncpg = "^0.30.0"
python-multipart = "^0.0.20"
bcrypt = "^4.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
black = "^24.3.0"
mypy = "^1.8.0"  # Also adding mypy for type checking
types-python-dateutil = "^2.8.19"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
